Walsh 06/06/2016 - Garbage Scripts
Starting to do script for a new way of making money which involves the
player buying a garbage truck (not sure about model or whatnot) and
going around the map to placed wheelie bins and collecting "trash".
They'd then go on to sell it at some guy somewhere who buys trash. I'm
at a point where I am stuck and don't know what to do next, so if
someone wants to either continue with it or help me at some point this
weekend or in the future that would be great.