Changed boltcutter speed to be slower/respawn timer to 4 minutes
Changed boltcutter because it was just too easy to go into a town and
break into
all the houses there in less than 5 minutes. This adds at least some
wait time to breaking down each door. The respawn timer was done because
EMS was having a struggle going and getting someone if they were on the
other side of the map.